+++ packages/eas-cli/src/graphql/mutations/KeystoreGenerationUrlMutation.ts
@@ -18,7 +18,8 @@ export const KeystoreGenerationUrlMutation = {
                 }
               }
             }
-          `
+          `,
+          {}
         )
         .toPromise()
     );